Simplified and fixed code for checking preferred contacts.
Now in case one contact of a given flavour is specified, it is not needed to
check it as preferred (it is set as preferred by default)

NB there is a problem in the model. Now we are sharing contacts among people, but the
flag "preferred" is meaningful to a person and not to another. change the model or
avoid contact sharing